Extract useOpenCloseable hook
